requirements.txt
boto3==1.9.251
botocore==1.12.251
dj-database-url==0.5.0
Django==3.1.8
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-storages==1.10.1
docutils==0.15.2
gunicorn==19.9.0
jmespath==0.9.4
Pillow==8.1.1
psycopg2==2.8.3
python-dateutil==2.8.0
pytz==2019.3
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.8
whitenoise==4.1.4
pytest==5.1.1
selenium==3.14.0

